admin 게시물 신고 리스트 테스트

+++ src/main/java/com/example/board/toyboard/DTO/PostReportDTO.java
@@ -14,9 +14,9 @@ public class PostReportDTO {
 
     private String nickname;
 
-    private int reposts;
+    private Long reposts;
 
-    public PostReportDTO(Long id, String title, String nickname, int reposts) {
+    public PostReportDTO(Long id, String title, String nickname, Long reposts) {
         this.id = id;
         this.title = title;
         this.nickname = nickname;
